Start adding Mocha tests for failures GraphQL API

+++ controllers/graphql/schema/globalTypes/rangeType.js
@@ -2,9 +2,9 @@ import { gql } from 'apollo-server-express';
 
 export default gql`
   input RangeInput {
-    maxLatitude: Int
-    minLatitude: Int
-    maxLongitude: Int
-    minLongitude: Int
+    maxLatitude: Float
+    minLatitude: Float
+    maxLongitude: Float
+    minLongitude: Float
   }
 `;
